fix(metrics): discover repos driven via git -C for commit hydration
Commit hydration derived its repo roots from process.cwd(), NR_AI_GIT_REPOS
and session cwds only. Work driven from one workspace but targeted at another
repo (`git -C <path> ...`) never changes cwd, so that repo was invisible to
hydration and the dashboard only ever listed the workspace's own repos.

Record the resolved target directory of every git command and union it into
the aggregator's cwds(), so a repo touched only via `git -C` is discovered.
gitCommandTargetDir moves into local-session-aggregator to avoid a circular
import (git-efficiency-tracker already imports RepoNameResolver from it).
